fix: users keep authorized when they deleted their own profiles
